LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-datetime-expected.txt

Tests if calendar picker works for datetime as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS selectedDate() is "2011-12-31"
PASS selectedDate() is "2012-01-01"
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("datetime").value is "2012-01-01T12:53:45Z"
PASS successfullyParsed is true

TEST COMPLETE
